physical-ai-robotics-textbook/backend/requirements.txt
fastapi==0.124.0
uvicorn==0.38.0
pydantic==2.12.5
pydantic-settings==2.12.0
google-generativeai==0.8.5
qdrant-client==1.16.1
sqlalchemy>=1.4.42,<2.0.0  # Compatible with databases
asyncpg==0.31.0
databases==0.8.0  # For async database support
python-multipart==0.0.20
python-dotenv==1.2.1
requests==2.32.5
tiktoken==0.12.0
python-socketio==5.15.0
httpx==0.28.1
pytest==9.0.1
pytest-asyncio==1.3.0
pytest-cov==7.0.0
alembic>=1.12.0  # For database migrations
python-frontmatter==1.1.0  # For parsing markdown with YAML frontmatter
markdown==3.7  # For markdown processing
beautifulsoup4==4.12.3  # For HTML parsing